requirements.txt
mcp[cli]==1.19.0
pydantic==2.12.3
python-dotenv==1.1.1
pydantic-settings==2.11.0
pymongo==4.10.1
